dara_llm/pyproject.toml
[tool.poetry]
name = "dara_llm"
authors = []
description = ""
version = "1.0.0"

[tool.poetry.dependencies]
python = ">=3.8.0, <3.12.0"
dara-core = ">=0.3.3, <1.0.0"
dara-components = ">=0.3.3, <1.0.0"
openai = "^0.27.8"
statsmodels = "^0.14.0"
scikit-learn = "^1.3.0"
